refactor: refactor social listening pages and components
